split state into state and io modules
