+++ sphinx/source/docs/reference/resources_embedding.rst
@@ -14,6 +14,15 @@ Resources and Embedding
 .. automodule:: bokeh.state
   :members:
 
+.. _bokeh.io:
+
+``bokeh.io``
+------------
+
+.. automodule:: bokeh.io
+  :members:
+
+
 .. _bokeh_dot_resources:
 
 ``bokeh.resources``
